requirements.txt
Package               Version
--------------------- -------
asgiref               3.3.1
Django                3.1.4
django-rest-framework 0.1.0
djangorestframework   3.12.2
pip                   20.3.1
psycopg2              2.8.6
pytz                  2020.4
setuptools            44.1.1
sqlparse              0.4.1

Django~=3.1.4
requests~=2.25.0